Do not allow try statements
