+++ src/compiler/index.ts
@@ -3,16 +3,22 @@ export * from "./await";
 export * from "./binary";
 export * from "./binding";
 export * from "./block";
+export * from "./break";
 export * from "./call";
 export * from "./class";
 export * from "./conditional";
+export * from "./continue";
+export * from "./do";
 export * from "./enum";
 export * from "./expression";
+export * from "./for";
+export * from "./forOf";
 export * from "./function";
 export * from "./identifier";
 export * from "./if";
 export * from "./indexed";
 export * from "./literal";
+export * from "./loop";
 export * from "./module";
 export * from "./namespace";
 export * from "./new";
@@ -25,16 +31,9 @@ export * from "./spread";
 export * from "./statement";
 export * from "./switch";
 export * from "./template";
-export * from "./try";
+export * from "./throw";
+export * from "./truthiness";
 export * from "./unary";
 export * from "./variable";
-export * from "./yield";
-
-export * from "./break";
-export * from "./continue";
-export * from "./do";
-export * from "./for";
-export * from "./forOf";
-export * from "./loop";
 export * from "./while";
-export * from "./truthiness";
+export * from "./yield";
