+++ tests/src/errors/try.spec.ts
@@ -1,0 +1,2 @@
+export {}
+try {} catch {}
